Reorganize entrypoint functions
