+++ crates/cgp-component-macro-lib/src/delegate_components/mod.rs
@@ -1,7 +1,4 @@
 pub mod define_struct;
-pub mod delegate;
 pub mod delegates_to;
 pub mod impl_delegate;
 pub mod merge_generics;
-
-pub use delegate::delegate_components;
