+++ crates/cgp-component-macro-lib/src/derive_component/mod.rs
@@ -12,5 +12,3 @@ pub mod replace_self_receiver;
 pub mod replace_self_type;
 pub mod signature_args;
 pub mod snake_case;
-
-pub use derive::derive_component;
