+++ crates/cgp-component-macro-lib/src/derive_context/derive.rs
@@ -1,45 +1,7 @@
-use proc_macro2::TokenStream;
 use quote::quote;
 use syn::{parse2, parse_quote, Ident, ItemImpl, ItemStruct, Path};
 
-use crate::derive_context::ContextSpec;
 use crate::parse::TypeGenerics;
-
-pub fn derive_context(attr: TokenStream, body: TokenStream) -> syn::Result<TokenStream> {
-    let context_spec: ContextSpec = syn::parse2(attr)?;
-
-    let context_struct: ItemStruct = syn::parse2(body)?;
-
-    let provider_name = &context_spec.provider_name;
-
-    let provider_struct: ItemStruct = parse_quote!( pub struct #provider_name; );
-
-    let has_components_impl: ItemImpl = derive_has_components(provider_name, &context_struct);
-
-    let base_derived = quote! {
-        #context_struct
-
-        #provider_struct
-
-        #has_components_impl
-    };
-
-    match &context_spec.preset {
-        Some(preset) => {
-            let (delegate_impl, is_provider_impl) =
-                derive_delegate_preset(provider_name, &preset.name, &preset.generics)?;
-
-            Ok(quote! {
-                #base_derived
-
-                #delegate_impl
-
-                #is_provider_impl
-            })
-        }
-        _ => Ok(base_derived),
-    }
-}
 
 pub fn derive_has_components(provider_name: &Ident, context_struct: &ItemStruct) -> ItemImpl {
     let context_name = &context_struct.ident;
