+++ crates/cgp-component-macro-lib/src/entrypoints/cgp_auto_getter.rs
@@ -1,0 +1,28 @@
+use proc_macro2::{Span, TokenStream};
+use quote::quote;
+use syn::{Error, Ident, ItemTrait};
+
+use crate::getter_component::blanket::derive_blanket_impl;
+use crate::getter_component::parse::parse_getter_fields;
+
+pub fn cgp_auto_getter(attr: TokenStream, body: TokenStream) -> syn::Result<TokenStream> {
+    if !attr.is_empty() {
+        return Err(Error::new(
+            Span::call_site(),
+            "#[derive_auto_getter] does not accept any attribute argument",
+        ));
+    }
+
+    let consumer_trait: ItemTrait = syn::parse2(body)?;
+
+    let context_type = Ident::new("__Context__", Span::call_site());
+
+    let fields = parse_getter_fields(&context_type, &consumer_trait)?;
+
+    let blanket_impl = derive_blanket_impl(&context_type, &consumer_trait, &fields);
+
+    Ok(quote! {
+        #consumer_trait
+        #blanket_impl
+    })
+}
