+++ crates/cgp-component-macro-lib/src/entrypoints/define_preset.rs
