+++ crates/cgp-component-macro-lib/src/entrypoints/delegate_components.rs
