+++ crates/cgp-component-macro-lib/src/for_each_replace.rs
@@ -83,16 +83,6 @@ pub fn handle_for_each_replace(tokens: TokenStream) -> syn::Result<TokenStream> 
     ))
 }
 
-pub fn handle_replace(tokens: TokenStream) -> syn::Result<TokenStream> {
-    let specs: ReplaceSpecs = syn::parse2(tokens)?;
-
-    let items: Punctuated<TokenStream, Comma> = specs.replacements.into_iter().collect();
-
-    let tokens = quote! { [ #items ] };
-
-    Ok(replace_stream(&specs.target_ident, &tokens, specs.body))
-}
-
 pub fn for_each_replace(
     target_ident: &Ident,
     replacements: &[TokenStream],
