+++ crates/cgp-component-macro-lib/src/getter_component/mod.rs
@@ -1,5 +1,4 @@
 pub mod blanket;
-pub mod derive;
 pub mod getter_field;
 pub mod parse;
 pub mod symbol;
