+++ crates/cgp-component-macro-lib/src/preset/mod.rs
@@ -1,4 +1,2 @@
-pub mod define_preset;
-pub use define_preset::define_preset;
 pub mod impl_is_preset;
 pub mod substitution_macro;
